--- pars/a
@@ -1,16 +1,0 @@
-222221111111111111111111111111222222222
-222210000000001100000000000001222222222
-222211110110000011100000000001222222222
-222210000000000000000000000111122222222
-111111111011000001110000000000001222222
-100000000011000001110111111111111222222
-111101111111110111000000100012222222222
-110000001101010111000000100012222222222
-100000000000000110000000010111222222222
-1000000000000000000P0000010101111111111
-110000011101011111101110000111222222222
-111101112111010121011110100011222222222
-111111112111111121111111111112222222222
-111111112111111121111111111112222222222
-111111112111111121111111111112222222222
-111111112111111121111111111112222222222
