+++ pars/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: abelahce <abelahce@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/06 08:06:57 by abelahce          #+#    #+#             */
-/*   Updated: 2023/03/10 02:30:52 by abelahce         ###   ########.fr       */
+/*   Updated: 2023/04/03 01:28:20 by abelahce         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ size_t	ft_strlen(const char	*str)
 
 	i = 0;
 	if (str == NULL)
-		return(0);
+		return (0);
 	while (str[i] != '\0')
 		i++;
 	return (i);
