+++ pars/libft/ft_strncmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abelahce <abelahce@student.42.fr>          +#+  +:+       +#+        */
+/*   By: abelahce <abelahce@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 09:43:06 by abelahce          #+#    #+#             */
-/*   Updated: 2021/11/29 03:58:30 by abelahce         ###   ########.fr       */
+/*   Updated: 2023/04/02 05:13:26 by abelahce         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
