+++ pars/map.cub
@@ -1,21 +1,22 @@
 WE  xpm/test2.xpm 
 NO  xpm/test4.xpm
-SO  xpm/test1.xpm
 EA  xpm/test3.xpm
 
+C 225,30,255
 F 255,100,1
-C 225,30,0
 
+
+SO xpm/test1.xpm
 
 
 11111111111111111111111111111111111111
 1000100000000W000000000000000000000001
 11111111111100000000000000000000000001
-           100011111111111111111100001
+           100111111111111111111000001
            10001                100001
            10001                100001
            10001                100001
            10001                100001
            100011111111111111111100001
-           100000000000000000000000001
-           111111111111111111111111111
+           1000000000000001000000000011
+           1111111111111111111111111111
